+++ services/elph/.objects
@@ -1,1 +1,2 @@
-objs = set_kernel.o elph_driver.o compute_elph_matq.o compute_dmats.o ep_f2c_bridge.o dvG_utils.o
+objs = elph_driver_cb.o set_kernel.o elph_driver.o ep_f2c_bridge.o dvG_utils.o compute_elph_matq.o compute_dmats.o \
+       elph_driver_cb2.o
