+++ services/preprocessor/preprocessor.h
@@ -9,7 +9,11 @@
 
 void ELPH_cli_parser(int argc, char* argv[], struct calc_details* calc_info);
 
-void create_ph_save_dir_pp_qe(const char* inp_file,const char* ph_path);
+#if defined _Y6_LETZ
+void create_ph_save_dir_pp_qe(const char* inp_file,char* ph_path,char* scf_path);
+#else
+void create_ph_save_dir_pp_qe(const char* inp_file,const char* ph_path, const char* scf_path);
+#endif
 
 void ELPH_print_version(void);
 void ELPH_print_help(void);
